chore: move bbun module
